+++ packages/date-adapters/src/__tests__/dayjs.spec.ts
@@ -3,7 +3,7 @@ import customParseFormat from "dayjs/plugin/customParseFormat";
 import timezone from "dayjs/plugin/timezone";
 import utc from "dayjs/plugin/utc";
 import { describe, expect, it } from "vitest";
-import { AdapterDayjs, DateDetailErrorEnum } from "../../date-adapters";
+import { AdapterDayjs, DateDetailErrorEnum } from "../index";
 
 // Extend dayjs with necessary plugins
 dayjs.extend(utc);
