+++ packages/date-adapters/src/index.ts
@@ -10,12 +10,9 @@
 // biome-ignore lint/complexity/noBannedTypes: type augmented by configured adapters
 export type DateFrameworkTypeMap = {};
 
-export * from "./types";
-export * from "./LocalizationProvider";
-
-// Supported date libraries but you can create your own by implementing `SaltDateAdapter`
 export * from "./date-fns";
 export * from "./dayjs";
 export * from "./luxon";
-/** Deprecated by maintainers since Sept 2020, consider an alternative OSS date library instead */
 export * from "./moment";
+
+export * from "./types";
