+++ packages/date-adapters/src/luxon/index.ts
@@ -7,9 +7,9 @@ import {
   type SaltDateAdapter,
   type TimeFields,
   type Timezone,
-} from "./types";
+} from "../types";
 
-declare module "./types" {
+declare module "../types" {
   interface DateFrameworkTypeMap {
     luxon: DateTime;
   }
