+++ packages/lab/src/__tests__/__e2e__/calendar/Calendar.multiselect.cy.tsx
@@ -1,14 +1,16 @@
+import type {
+  DateFrameworkType,
+  SaltDateAdapter,
+} from "@salt-ds/date-adapters";
+import { AdapterDateFns } from "@salt-ds/date-adapters";
+import { AdapterDayjs } from "@salt-ds/date-adapters";
+import { AdapterLuxon } from "@salt-ds/date-adapters";
+import { AdapterMoment } from "@salt-ds/date-adapters";
 import {
-  AdapterDateFns,
-  AdapterDayjs,
-  AdapterLuxon,
-  AdapterMoment,
   Calendar,
   CalendarGrid,
   CalendarNavigation,
   CalendarWeekHeader,
-  type DateFrameworkType,
-  type SaltDateAdapter,
 } from "@salt-ds/lab";
 
 // Initialize adapters
