+++ packages/lab/src/__tests__/__e2e__/date-input/DateInputSingle.cy.tsx
@@ -1,14 +1,15 @@
 import {
-  AdapterDateFns,
-  AdapterDayjs,
-  AdapterLuxon,
-  AdapterMoment,
   DateDetailErrorEnum,
   type DateFrameworkType,
-  DateInputSingle,
   type ParserResult,
   type SaltDateAdapter,
-} from "@salt-ds/lab";
+} from "@salt-ds/date-adapters";
+import { AdapterDateFns } from "@salt-ds/date-adapters";
+import { AdapterDayjs } from "@salt-ds/date-adapters";
+import { AdapterLuxon } from "@salt-ds/date-adapters";
+import { AdapterMoment } from "@salt-ds/date-adapters";
+import { DateInputSingle } from "@salt-ds/lab";
+
 import { es as dateFnsEs } from "date-fns/locale";
 import { type SyntheticEvent, useState } from "react";
 import "moment/dist/locale/es";
