+++ packages/lab/src/calendar/internal/CalendarMonth.tsx
@@ -9,9 +9,10 @@ import {
 import { CalendarDay, type CalendarDayProps } from "./CalendarDay";
 import { generateVisibleDays } from "./utils";
 
+import type { DateFrameworkType } from "@salt-ds/date-adapters";
 import { useComponentCssInjection } from "@salt-ds/styles";
 import { useWindow } from "@salt-ds/window";
-import { type DateFrameworkType, useLocalization } from "../../date-adapters";
+import { useLocalization } from "../../localization-provider";
 import { useCalendarContext } from "./CalendarContext";
 import calendarMonthCss from "./CalendarMonth.css";
