+++ packages/lab/src/date-picker/DatePickerRangePanel.tsx
@@ -8,6 +8,10 @@ import {
   makePrefixer,
   useControlled,
 } from "@salt-ds/core";
+import type {
+  DateFrameworkType,
+  SaltDateAdapter,
+} from "@salt-ds/date-adapters";
 import { useComponentCssInjection } from "@salt-ds/styles";
 import { useWindow } from "@salt-ds/window";
 import clsx from "clsx";
@@ -32,11 +36,7 @@ import {
   type DateRangeSelection,
   type UseCalendarSelectionRangeProps,
 } from "../calendar";
-import {
-  type DateFrameworkType,
-  type SaltDateAdapter,
-  useLocalization,
-} from "../date-adapters";
+import { useLocalization } from "../localization-provider";
 import { useDatePickerContext } from "./DatePickerContext";
 import datePickerPanelCss from "./DatePickerPanel.css";
