+++ packages/lab/src/localization-provider/LocalizationProvider.tsx
@@ -1,5 +1,8 @@
+import type {
+  DateFrameworkType,
+  SaltDateAdapter,
+} from "@salt-ds/date-adapters";
 import { createContext, useContext, useMemo } from "react";
-import type { DateFrameworkType, SaltDateAdapter } from "./types";
 
 const MIN_DATE = "1900-01-01T00:00:00.000";
 const MAX_DATE = "2099-12-31T00:00:00.000";
