+++ packages/lab/src/localization-provider/index.ts
@@ -1,0 +1,1 @@
+export * from "./LocalizationProvider";
